backend/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
neo4j==4.2.0
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pymongo==3.11.2
Werkzeug==1.0.1
